requirements.txt
blinker==1.6.2
click==8.1.6
Flask==2.3.2
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
python-docx==0.8.11
Werkzeug==2.3.7
